\name{getdata}
\alias{getdata}
\title{Get Soprano Quote Data from Web}
\usage{
getdata()
}
\description{
Returns a data frame with Sopranos Quotes
Sourced from static website https://www.thedad.com/the-sopranos-quotes/
Quotes partitioned by Tony, Soprano family, and everybody else
Not exhaustive of all quotes in series.
}
\examples{
df <- getdata()
}
