\name{getquote}
\alias{getquote}
\title{Get Random Sopranos Quote}
\usage{
getquote()
}
\description{
Prints a random Sopranos quote based on desired speaker
(Tony, Soprano Family, Everybody Else).
Defaults to any.
}
\examples{
getquote()
getquote(person='tony')
getquote(person='family')
getquote(person='everybody')
}
